refactor(about): share organization model
